requirements/base.txt
Django==2.0.4
psycopg2==2.7.4
django-model-utils==3.1.1
python-dateutil==2.6.1
djangorestframework==3.8.2
pycountry==18.2.23
djangorestframework-gis==0.12
